requirements.txt
streamlit==1.29.0
requests==2.31.0
plotly==5.18.0
pandas==2.1.4
deep-translator==1.11.4
langdetect==1.0.9
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
beautifulsoup4==4.12.2
